rename ServerList.py to server_list.py
