+++ server/src/server_browser_backend/__init__.py
@@ -1,3 +1,4 @@
 from .models import *
 from .dict_util import *
 from .main import *
+from .server_list import *
